+++ packages/sandbox/src/tools/index.ts
@@ -5,7 +5,6 @@ import { shellTool } from "./shell.js";
 import { stopTool } from "./stop.js";
 import { cleanTool } from "./clean.js";
 import { sandboxTool } from "./sandbox.js";
-import { heartbeatTool } from "./heartbeat.js";
 import { worktreeTool } from "./worktree.js";
 import { onboardTool } from "./onboard.js";
 import { portsTool } from "./ports.js";
@@ -20,7 +19,6 @@ export const sandboxTools: ToolDefinition[] = [
   shellTool,
   stopTool,
   cleanTool,
-  heartbeatTool,
   worktreeTool,
   onboardTool,
   portsTool,
@@ -36,7 +34,6 @@ export {
   stopTool,
   cleanTool,
   sandboxTool,
-  heartbeatTool,
   worktreeTool,
   onboardTool,
   portsTool,